+++ src/main/java/com/CuteNekoDragon/Core/common/data/SVOItemPrices.java
@@ -66,8 +66,12 @@ public class SVOItemPrices {
 
         // Special Items
         PriceUtil.setPrice(SVOItems.GALAXY_SOUL, 5000, PriceUtil.PriceType.NONE);
+        PriceUtil.setPrice(SVOItems.DRAGON_TOOTH, 500, PriceUtil.PriceType.NONE);
+        PriceUtil.setPrice(SVOItems.CINDER_SHARD, 50, PriceUtil.PriceType.NONE);
 
         // Ingots and Blacksmith Items
         PriceUtil.setPrice(SVOItems.IRIDIUM_INGOT, 1000, PriceUtil.PriceType.BLACKSMITH);
+        PriceUtil.setPrice(SVOItems.REFINED_QUARTZ, 50, PriceUtil.PriceType.BLACKSMITH);
+        PriceUtil.setPrice(SVOItems.RADIOACTIVE_INGOT, 3000, PriceUtil.PriceType.BLACKSMITH);
     }
 }